Added abstract AssetProcessingService and SketchAssetProcessor
