+++ lib/input/helper/asset_processing_service.dart
@@ -1,0 +1,5 @@
+abstract class AssetProcessingService {
+  dynamic processImage(String uuid);
+
+  Future<void> processRootElements(List<String> uuids);
+}
