+++ lib/input/sketch/entities/layers/polygon.dart
@@ -1,11 +1,12 @@
 import 'package:json_annotation/json_annotation.dart';
+import 'package:parabeac_core/input/helper/asset_processing_service.dart';
 import 'package:parabeac_core/input/sketch/entities/abstract_sketch_node_factory.dart';
 import 'package:parabeac_core/input/sketch/entities/layers/abstract_layer.dart';
 import 'package:parabeac_core/input/sketch/entities/layers/abstract_shape_layer.dart';
 import 'package:parabeac_core/input/sketch/entities/layers/flow.dart';
 import 'package:parabeac_core/input/sketch/entities/objects/frame.dart';
 import 'package:parabeac_core/input/sketch/entities/style/style.dart';
-import 'package:parabeac_core/input/sketch/helper/svg_png_convertion.dart';
+import 'package:parabeac_core/input/sketch/helper/sketch_asset_processor.dart';
 import 'package:parabeac_core/interpret_and_optimize/entities/inherited_polygon.dart';
 import 'package:parabeac_core/interpret_and_optimize/entities/subclasses/pb_intermediate_node.dart';
 import 'package:parabeac_core/interpret_and_optimize/helpers/pb_context.dart';
@@ -117,8 +118,8 @@ class Polygon extends AbstractShapeLayer implements SketchNodeFactory {
 
   @override
   Future<PBIntermediateNode> interpretNode(PBContext currentContext) async {
-    var image = await convertImage(
-        UUID, boundaryRectangle.width, boundaryRectangle.height);
+    var image = await SketchAssetProcessor()
+        .processImage(UUID, boundaryRectangle.width, boundaryRectangle.height);
 
     return Future.value(InheritedPolygon(this, name,
         currentContext: currentContext, image: image));
